Fixed compiler warnings: 103 unnecessary casts (CASTOR-1393)

+++ trunk/castor/src/main/org/exolab/castor/types/GMonth.java
@@ -123,7 +123,7 @@ public class GMonth extends GMonthDay {
         if ( isUTC() ) {
             SimpleTimeZone timeZone = new SimpleTimeZone(0,"UTC");
             int offset = 0;
-            offset = (int) ( (this.getZoneMinute() + this.getZoneHour()*60)*60*1000);
+            offset = ( (this.getZoneMinute() + this.getZoneHour()*60)*60*1000);
             offset = isZoneNegative() ? -offset : offset;
             timeZone.setRawOffset(offset);
             timeZone.setID(TimeZone.getAvailableIDs(offset)[0]);
